Move error to separate class
